requirements.txt
asgiref==3.7.2
Django==4.2.4
django-autoslug==1.9.9
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
